Add create_policy function to create q_net and optimizer
The create_policy function takes in an environment and creates a q_net
and an Adam optimizer.
The q_net is a simple example network with embedding, gru and output layer.
The input is a single string and the output a vector with a value for each possible action.
The main method performs a single call of the model for testing.